+++ main.py
@@ -1,15 +1,16 @@
 import sys
 
+from PyQt5.QtCore import Qt
 from PyQt5.QtWidgets import QApplication
 
-from main_view import MainView
+from src.main_view import MainView
 
 
 class App(QApplication):
     def __init__(self, sys_argv):
         super(App, self).__init__(sys_argv)
         self.main_view = MainView()
-        self.main_view.resize(1920, 1000)
+        self.main_view.setWindowState(Qt.WindowMaximized)
         self.main_view.show()
