+++ src/qt_dialogs/selectDialog_view.py
@@ -2,8 +2,8 @@ from PyQt5 import QtWidgets
 from PyQt5.QtCore import Qt, pyqtSlot
 from PyQt5.QtWidgets import QDialog, QTableWidgetItem, QAbstractItemView
 
-from custom_signals import CustomSignals
-from qt_ui.selectDialog_ui import Ui_SelectDialog
+from src.custom_signals import CustomSignals
+from src.qt_ui.selectDialog_ui import Ui_SelectDialog
 
 
 class SelectDialog(QDialog):
