+++ src/qt_ui/dynamicsDialog_ui.py
